// cruscotto/src/app/modules/domande-psr-strutturali/models/domanda-psr-strutturale.ts
export class DomandaPsrStrutturale {
    idBando: number;
    codiceEnte: string;
    cuaa: string;
    descrizioneBando: string;
    descrizioneNormativa: string;
    idProgetto: number;
    codiceStatoProgetto: string;
    statoProgetto: string;
    codiceOperazione: string;
    protocolloProgetto: string;
    dataProgetto: string;
    dataTermineRendicontazione: Date;
}