// cruscotto/src/app/modules/domande-psr-strutturali/models/dettaglio-domanda-psr-strutturale.ts
export class DettaglioPSRStrutturale {
    idProgetto: number;
    descrizioneCodificaInvestimento: string;
    descrizioneDettaglioInvestimenti: string;
    descrizionePianoInvestimenti: string;
    descrizioneSettoriProduttivi: string;
    costoInvestimento: number;
    speseTecniche: number;
    percentualeRichiesto: number;
    contributoRichiesto: number;
  }