// cruscotto/src/app/modules/domande-psr-strutturali/models/investimento.ts
export class Investimento {
  idInvestimento: number;
  idProgetto: number;
  descrizione: string;
  costoInvestimento: number;
  speseGenerali: number;
  contributoRichiesto: number;
  quotaContributoRichiesto: number;
  idVariante: number;
  idInvestimentoOriginale: number;
  descrizioneCodificaInvestimento: string;
  descrizioneDettaglioInvestimenti: string;
  descrizioneSettoriProduttivi: string;
}